Fix Laravel 9 authentication and permission system migration
- Fix AuthController redirectPath() method missing
- Add LevelMiddleware for legacy level:X route middleware
- Update User model with spatie/laravel-permission compatibility methods
- Fix auth configuration for Laravel 9
- Update permission tables migration for existing schema compatibility
- Add guard_name column migration for roles table
- Recreate all seeders for spatie/laravel-permission
- Fix role names in views: admin->AOE Admin, reseller->Reseller, client->Client Admin
- Update sidebar, navigation, and dashboard views with correct role names

+++ database/migrations/2025_08_21_142840_create_permission_tables.php
@@ -27,7 +27,7 @@ class CreatePermissionTables extends Migration
 
         if (!Schema::hasTable($tableNames['permissions'])) {
             Schema::create($tableNames['permissions'], function (Blueprint $table) {
-                $table->bigIncrements('id'); // permission id
+                $table->increments('id'); // permission id
                 $table->string('name');       // For MySQL 8.0 use string('name', 125);
                 $table->string('guard_name'); // For MySQL 8.0 use string('guard_name', 125);
                 $table->timestamps();
@@ -38,9 +38,9 @@ class CreatePermissionTables extends Migration
 
         if (!Schema::hasTable($tableNames['roles'])) {
             Schema::create($tableNames['roles'], function (Blueprint $table) use ($teams, $columnNames) {
-                $table->bigIncrements('id'); // role id
+                $table->increments('id'); // role id
                 if ($teams || config('permission.testing')) { // permission.testing is a fix for sqlite testing
-                    $table->unsignedBigInteger($columnNames['team_foreign_key'])->nullable();
+                    $table->unsignedInteger($columnNames['team_foreign_key'])->nullable();
                     $table->index($columnNames['team_foreign_key'], 'roles_team_foreign_key_index');
                 }
                 $table->string('name');       // For MySQL 8.0 use string('name', 125);
@@ -56,10 +56,10 @@ class CreatePermissionTables extends Migration
 
         if (!Schema::hasTable($tableNames['model_has_permissions'])) {
             Schema::create($tableNames['model_has_permissions'], function (Blueprint $table) use ($tableNames, $columnNames, $teams) {
-                $table->unsignedBigInteger(PermissionRegistrar::$pivotPermission);
+                $table->unsignedInteger(PermissionRegistrar::$pivotPermission);
 
                 $table->string('model_type');
-                $table->unsignedBigInteger($columnNames['model_morph_key']);
+                $table->unsignedInteger($columnNames['model_morph_key']);
                 $table->index([$columnNames['model_morph_key'], 'model_type'], 'model_has_permissions_model_id_model_type_index');
 
                 $table->foreign(PermissionRegistrar::$pivotPermission)
@@ -67,7 +67,7 @@ class CreatePermissionTables extends Migration
                     ->on($tableNames['permissions'])
                     ->onDelete('cascade');
                 if ($teams) {
-                    $table->unsignedBigInteger($columnNames['team_foreign_key']);
+                    $table->unsignedInteger($columnNames['team_foreign_key']);
                     $table->index($columnNames['team_foreign_key'], 'model_has_permissions_team_foreign_key_index');
 
                     $table->primary([$columnNames['team_foreign_key'], PermissionRegistrar::$pivotPermission, $columnNames['model_morph_key'], 'model_type'],
@@ -82,10 +82,10 @@ class CreatePermissionTables extends Migration
 
         if (!Schema::hasTable($tableNames['model_has_roles'])) {
             Schema::create($tableNames['model_has_roles'], function (Blueprint $table) use ($tableNames, $columnNames, $teams) {
-                $table->unsignedBigInteger(PermissionRegistrar::$pivotRole);
+                $table->unsignedInteger(PermissionRegistrar::$pivotRole);
 
                 $table->string('model_type');
-                $table->unsignedBigInteger($columnNames['model_morph_key']);
+                $table->unsignedInteger($columnNames['model_morph_key']);
                 $table->index([$columnNames['model_morph_key'], 'model_type'], 'model_has_roles_model_id_model_type_index');
 
                 $table->foreign(PermissionRegistrar::$pivotRole)
@@ -93,7 +93,7 @@ class CreatePermissionTables extends Migration
                     ->on($tableNames['roles'])
                     ->onDelete('cascade');
                 if ($teams) {
-                    $table->unsignedBigInteger($columnNames['team_foreign_key']);
+                    $table->unsignedInteger($columnNames['team_foreign_key']);
                     $table->index($columnNames['team_foreign_key'], 'model_has_roles_team_foreign_key_index');
 
                     $table->primary([$columnNames['team_foreign_key'], PermissionRegistrar::$pivotRole, $columnNames['model_morph_key'], 'model_type'],
@@ -107,8 +107,8 @@ class CreatePermissionTables extends Migration
 
         if (!Schema::hasTable($tableNames['role_has_permissions'])) {
             Schema::create($tableNames['role_has_permissions'], function (Blueprint $table) use ($tableNames) {
-                $table->unsignedBigInteger(PermissionRegistrar::$pivotPermission);
-                $table->unsignedBigInteger(PermissionRegistrar::$pivotRole);
+                $table->unsignedInteger(PermissionRegistrar::$pivotPermission);
+                $table->unsignedInteger(PermissionRegistrar::$pivotRole);
 
                 $table->foreign(PermissionRegistrar::$pivotPermission)
                     ->references('id') // permission id
